+++ Assets/Scripts/Serializable/MonsterData.cs
@@ -1,5 +1,4 @@
 using Newtonsoft.Json;
-using System.Numerics;
 using System.Collections.Generic;
 
 public class MonsterDataList : IDataList<MonsterData>
@@ -15,8 +14,8 @@ public class MonsterData
     public string name;                 // 몬스터 명칭
     public string spriteName;           // 이미지 이름
 
-    public BigInteger maxHp;            // 체력
-    public BigInteger Attack;           // 공격력
+    public double maxHp;                // 체력
+    public double Attack;               // 공격력
 
     public MonsterType type;            // 몬스터 타입
     public SkillElement weakElement;    // 약점 속성
